# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Budget Tracker API
  version: 0.0.1
paths:
  /api/v1/jwt/issue:
   post:
      summary: "Create a new JWT token"
      description: "Create a new token to sign every request to the platform"
      tags:
      - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JWTUser'
            example:
              login: "vsantos"
              password: "randompass"
      parameters:
        - in: header
          name: Content-Type
          required: true
          schema:
            type: string
            example: application/json
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              example:
                type: "bearer"
                refresh: ""
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE3NTg0OTIxMzcsImlhdCI6MTc1ODQ5MTgzNywibmFtZSI6InZzYW50b3MiLCJzdWIiOiJ0ZXN0In0.jXxP85bOfYSSgTSGSz9jJRAnAeZrz0VVZ4RInT0f_aM"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              example:
                message: "required login and password to request a new token"
  /api/v1/users:    
    post:
      summary: "Create an user"
      description: "Create a new user to the platform"
      tags:
      - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              login: "vsantos"
              firstname: "Victor"
              lastname: "Santos"
              email: "vsnatos.py@gmail.com"
              password: "randompass"
      parameters:
        - in: header
          name: Content-Type
          required: true
          schema:
            type: string
            example: application/json
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                message: "created user 'vsantos'"
                id: "68d066e8a4b62b805177aa91"
        '409':
          description: Conflict
          content:
            application/json:
              example:
                message: "could not create user"
                details: "user or email already registered"
  /api/v1/users/{id}:    
    get:
      summary: "Get details from an user"
      description: "based on the user_id, get the details from a single user"
      tags:
      - Users
      parameters:
        - name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "68d066e8a4b62b805177aa91"
          description: user ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "687baad049572fb8c4e305f9"
                login: "vsantos"
                firstname: "Victor"
                lastname: "Santos"
                email: "vsantos.py@gmail.com"
                password: "<sensitive>"
                created_at: "2025-07-19T14:25:20.996Z"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
              example:
                message: "could not find user"
                id: "687baad049572fb8c4e305f1"
        '500':
          description: User not found
          content:
            application/json:
              schema:
                type: object
              example:
                message: "not able to get ID format"
  /api/v1/cards:
    post:
      summary: Create a card
      description: Endpoint to create a new card for a given owner_id
      tags:
      - Cards
      parameters:
        - in: header
          name: Content-Type
          required: true
          schema:
            type: string
            example: application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
            example:
              owner_id: "687baad049572fb8c4e305f3"
              alias: "Platinum"
              type: "Debit"
              network: "VISA"
              bank: "Citibank"
              color: "#fff"
              last_digits: 1234
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
              example:
                message: "card created"
                status_code: 201
                card:
                  id: "68d036ba865f1ea380a77fa9"
                  owner_id: "687baad049572fb8c4e305f3"
                  alias: "Platinum"
                  type: "Debit"
                  network: "VISA"
                  bank: "Citibank"
                  color: "#fff"
                  last_digits: 1234
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
              example:
                message: "could not create card"
                details: "card already registered with the 'last 4 digits'"
                status_code: 409
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
              example:
                message: "could not create card"
                details: "internal error when posting card"
                status_code: 500
  /api/v1/cards/{id}:    
    get:
      summary: "Get details from a card"
      description: "based on the card_id, get the details from a single card"
      tags:
      - Cards
      parameters:
        - in: header
          name: Content-Type
          required: true
          schema:
            type: string
            example: application/json
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "68d066e8a4b62b805177aa91"
          description: card ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
              example:
                id: "68d036ba865f1ea380a77fa9"
                owner_id: "687baad049572fb8c4e305f3"
                alias: "Platinum"
                type: "Debit"
                network: "VISA"
                bank: "Citibank"
                color: "#fff"
                last_digits: 1234
                created_at: "2025-09-21T17:32:49.61Z"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
              example:
                message: "could not find card"
                details: "card id '68c6246ba4bd976c14e03143' not found"
                status_code: 404
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
              example:
                message: "error when fetching card's details"
                details: "not able to get ID format"
    delete:
      summary: Delete card
      tags:
      - Cards
      parameters:
        - in: header
          name: Content-Type
          required: true
          schema:
            type: string
            example: application/json
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "68d066e8a4b62b805177aa91"
          description: card ID
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
              example:
                message: "card deleted"
                id: "68c5a209829c95bee34722fb"
                status_code: 200
        '404':
          description: Not found. In case of trying to delete a non-existent card
          content:
            application/json:
              schema:
                type: object
              example:
                message: "card not deleted"
                details: "no cards were deleted from given card ID '68d03563239a30e8335f43a4'"
                status_code: 404
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
              example:
                message: "error when fetching card's details"
                details: "not able to get ID format"
components:
  schemas:
    JWTUser:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    Token:
      type: object
      properties:
        type:
          type: string
        refresh:
          type: string
        token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        password:
          type: string
        created_at:
          type: string
    Card:
      type: object
      properties:
        owner_id:
          type: string
        alias:
          type: string
        type:
          type: string
        network:
          type: string
        bank:
          type: string
        color:
          type: string
        last_digits:
          type: integer